FocusOnTop

Description

LayoutModifier that pulls the focused Window to the front of the screen, while preserving the relative order of the rest of the Windows.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import FocusOnTop

Then edit your layoutHook by adding focusOnTop:

myLayout = focusOnTop (Tall 1 (3/100) (1/2)) ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

https://hackage.haskell.org/package/xmonad-contrib-0.17.0/docs/XMonad-Doc-Extending.html#Editing_the_layout_hook

This module works especially nicely with SnapLayout:

import SnapLayout
import FocusOnTop

myLayout = focusOnTop snapLayout ||| Full ||| etc..
  where
    snapLayout :: SnapLayout a
    snapLayout = def
main = xmonad def { layoutHook = myLayout }

focusOnTop :: l Window -> ModifiedLayout FocusOnTop l Window

focusOnTop applies the FocusOnTop LayoutModifier on the provided Layout, to pull the focused Window to the front of the screen, while preserving the relative order of the rest of the Windows.