FocusOnTopListOps

Description

A small collection of helper List operations that is primarily intended for use with the FocusOnTop module. They are, however, generic, and therefore could be used elsewhere, in theory.

Documentation

remove :: (a -> b -> Bool) -> [a] -> b -> ([a], Maybe a)

remove takes an equality predicate of a and b, a list of a, and a b, and removes the first occurrence of the b from the list of a, and returns it (if applicable). This is similar to Data.List's deleteBy, except that it also returns the deleted element, if any.

removeAssoc :: Eq a => [(a, b)] -> a -> ([(a, b)], Maybe (a, b))

removeAssoc is a specialization of remove for an association list.

reverseConcat :: [a] -> [a] -> [a]

reverseConcat reverses the first list and then concatenates it with the second.

maybeCons :: ([a], Maybe a) -> [a]

maybeCons will cons the second element of the tuple to the first, if it exists. Otherwise it returns the first element as-is.

reorder :: Eq a => [(a, b)] -> [a] -> [(a, b)]

reorder takes an association list, l, and a list of keys to the association list, o. The latter represents the desired ordering of the former. Any elements from l that match keys in o will be shuffled to match the relative ordering of o, and any elements in l that are not found in o will maintain their relative order. The matched elements will appear after the unmatched elements in the result.