SnapLayout

Description

LayoutClass that lets the user snap windows to particular edges or corners of the screen, at various fractions of screen width and/or height.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import SnapLayout

Then edit your layoutHook by adding the SnapLayout:

myLayout = snapLayout ||| Full ||| etc..
  where
    snapLayout :: SnapLayout a
    snapLayout = def
main = xmonad def { layoutHook = myLayout }

Or, to customize the fine adjustment increment:

myLayout = snapLayout ||| Full ||| etc..
  where
    snapLayout :: SnapLayout a
    snapLayout = def { adjustmentDelta = 3/100 }
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

https://hackage.haskell.org/package/xmonad-contrib-0.17.0/docs/XMonad-Doc-Extending.html#Editing_the_layout_hook

In the key-bindings, do something like:

, ((modm .|. controlMask, xK_KP_End      ), withFocused (sendMessage . Snap BottomLeft))
, ((modm .|. controlMask, xK_KP_Down     ), withFocused (sendMessage . Snap Bottom))
, ((modm .|. controlMask, xK_KP_Page_Down), withFocused (sendMessage . Snap BottomRight))
, ((modm .|. controlMask, xK_KP_Left     ), withFocused (sendMessage . Snap SnapLayout.Left))
, ((modm .|. controlMask, xK_KP_Right    ), withFocused (sendMessage . Snap SnapLayout.Right))
, ((modm .|. controlMask, xK_KP_Home     ), withFocused (sendMessage . Snap TopLeft))
, ((modm .|. controlMask, xK_KP_Up       ), withFocused (sendMessage . Snap Top))
, ((modm .|. controlMask, xK_KP_Page_Up  ), withFocused (sendMessage . Snap TopRight))
, ((modm .|. controlMask, xK_KP_Insert   ), withFocused (sendMessage . Unsnap))

, ((modm .|. controlMask .|. shiftMask, xK_KP_End),
      withFocused (sendMessage . FineAdjustmentMessage BottomLeftAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Down),
      withFocused (sendMessage . FineAdjustmentMessage BottomAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Page_Down),
      withFocused (sendMessage . FineAdjustmentMessage BottomRightAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Left),
      withFocused (sendMessage . FineAdjustmentMessage LeftAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Right),
      withFocused (sendMessage . FineAdjustmentMessage RightAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Home),
      withFocused (sendMessage . FineAdjustmentMessage TopLeftAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Up),
      withFocused (sendMessage . FineAdjustmentMessage TopAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Page_Up),
      withFocused (sendMessage . FineAdjustmentMessage TopRightAdjustment))
, ((modm .|. controlMask .|. shiftMask, xK_KP_Insert), withFocused (sendMessage . Unadjust))
...

For detailed instruction on editing the key binding see:

https://hackage.haskell.org/package/xmonad-contrib-0.17.0/docs/XMonad-Doc-Extending.html#Editing_key_bindings

data SnapLayout a

Constructors

SnapLayout

Fields

snapped :: !(Map Window FullLoc)

Map used internally to keep track of Window states (default: Map.empty)

adjustmentDelta :: Rational

Percent of screen to adjust by in response to each FineAdjustmentMessage (default: 3/100)

data SnapLoc

Constructors

Top

Bottom

Left

Right

TopLeft

TopRight

BottomLeft

BottomRight

data Snap

Constructors

Snap SnapLoc Window

data FineAdjustmentMessage

Constructors

FineAdjustmentMessage FineAdjustmentDirection Window

data FineAdjustmentDirection

Constructors

TopAdjustment

BottomAdjustment

LeftAdjustment

RightAdjustment

TopLeftAdjustment

TopRightAdjustment

BottomLeftAdjustment

BottomRightAdjustment

data Unsnap

Constructors

Unsnap Window

data Unadjust

Constructors

Unadjust Window